open-powerlifting-api/requirements.txt
pandas
psycopg2-binary
requests
fastapi
decimal
os
pydantic
python-dotenv
zipfile
logging
datetime
json